requirements.txt
TTS
discord.py
openai
requests-html
bs4
urllib3
python-dotenv
yt-dlp
PyNaCl
transformers>=4.30
scikit_learn>=1.1
pandas>=1.0
matplotlib>3.7